% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{package}
\name{alpine-package}
\alias{alpine-package}
\title{alpine: bias corrected transcript abundance estimation}
\description{
alpine is a package for estimating and visualizing many forms of sample-specific
biases that can arise in RNA-seq, including fragment length
distribution, positional bias on the transcript, read
start bias (random hexamer priming), and fragment GC content
(amplification). It also offers bias-corrected estimates of
transcript abundance (FPKM). It is currently designed for
un-stranded paired-end RNA-seq data.
}
\details{
See the package vignette for a detailed workflow.

The main functions in this package are:
\enumerate{
\item \link{buildFragtypes} - build out features for fragment types from exons of a single gene (GRanges)
\item \link{fitBiasModels} - fit parameters for one or more bias models over a set of ~100 medium to highly expressed single isoform genes (GRangesList)
\item \link{estimateAbundance} - given a set of genome alignments (BAM files) and a set of isoforms of a gene (GRangesList), estimate the transcript abundances for these isoforms (FPKM) for various bias models
\item \link{extractAlpine} - given a list of output from \code{estimateAbundance}, compile an FPKM matrix across transcripts and samples
\item \link{predictCoverage} - given the exons of a single gene (GRanges) predict the coverage for a set of samples given fitted bias parameters and compute the observed coverage
}

Some helper functions for preparing gene objects:
\enumerate{
\item \link{splitGenesAcrossChroms} - split apart "genes" where isoforms are on different chromosomes
\item \link{splitLongGenes} - split apart "genes" which cover a suspiciously large range, e.g. 1 Mb
\item \link{mergeGenes} - merge overlapping isoforms into new "genes"
}

Some other assorted helper functions:
\enumerate{
\item \link{normalizeDESeq} - an across-sample normalization for FPKM matrices
\item \link{getFragmentWidths} - return a vector estimated fragment lengths given a set of exons for a single gene (GRanges) and a BAM file
\item \link{getReadLength} - return the read length of the first read across BAM files
}

The plotting functions are:
\enumerate{
\item \link{plotGC} - plot the fragment GC bias curves
\item \link{plotFragLen} - plot the framgent length distributions
\item \link{plotRelPos} - plot the positional bias (5' to 3')
\item \link{plotOrder0}, \link{plotOrder1}, \link{plotOrder2} - plot the read start bias terms
\item \link{plotGRL} - a simple function for visualizing GRangesList objects
}
}
\author{
Michael Love
}
\references{
Michael I Love, John B Hogenesch, Rafael A Irizarry:
Modeling of RNA-seq fragment sequence bias reduces
systematic errors in transcript abundance estimation
Posted to bioRxiv August 2015, \url{http://biorxiv.org/content/early/2015/08/28/025767}
}
\keyword{package}

