% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{buildFragtypes}
\alias{buildFragtypes}
\title{Build fragment types from exons}
\usage{
buildFragtypes(exons, genome, readlength, minsize, maxsize, gc = TRUE,
  gc.str = TRUE, vlmm = TRUE)
}
\arguments{
\item{exons}{a GRanges object with the exons for a single transcript}

\item{genome}{a BSgenome object}

\item{readlength}{the length of the reads. This doesn't necessarily
have to be exact (+/- 1 bp is acceptable)}

\item{minsize}{the minimum fragment length to model. The interval between
\code{minsize} and \code{maxsize} should contain the at least the
central 95 percent of the fragment length distribution across samples}

\item{maxsize}{the maximum fragment length to model}

\item{gc}{logical, whether to calculate the fragment GC content}

\item{gc.str}{logical, whether to look for presence of
stretches of very high GC within fragments}

\item{vlmm}{logical, whether to calculate the Cufflinks Variable Length
Markov Model (VLMM) for read start bias}
}
\value{
a DataFrame with bias features (columns) for all
potential fragments (rows)
}
\description{
This function constructs a DataFrame of fragment features used for
bias modeling, with one row for every potential fragment type that could
arise from a transcript. The output of this function is used by
\link{fitBiasModels}, and this function is used inside \link{estimateTheta}
in order to model the bias affecting different fragments across isoforms
of a gene.
}
\examples{

library(GenomicRanges)
library(BSgenome.Hsapiens.NCBI.GRCh38)
data(ebtfit)
readlength <- 100
minsize <- 125 # see vignette how to choose
maxsize <- 175 # see vignette how to choose
fragtypes <- buildFragtypes(ebt.fit[["ENST00000624447"]],
                            Hsapiens, readlength,
                            minsize, maxsize)

}

