% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bias.R
\name{fitBiasModels}
\alias{fitBiasModels}
\title{Fit bias models over single-isoform genes}
\usage{
fitBiasModels(genes, bamfile, fragtypes, genome, models, readlength, minsize,
  maxsize, speedglm = TRUE)
}
\arguments{
\item{genes}{a GRangesList with the exons of different genes}

\item{bamfile}{a character string pointing to an indexed BAM file}

\item{fragtypes}{the output of \link{buildFragtypes}. must contain
the potential fragment types for the genes named in \code{genes}}

\item{genome}{a BSGenome object}

\item{models}{a list of character strings or formula describing the bias models, see vignette}

\item{readlength}{the read length}

\item{minsize}{the minimum fragment length to model}

\item{maxsize}{the maximum fragment length to model}

\item{speedglm}{logical, whether to use speedglm to estimate the coefficients.
Default is TRUE.}
}
\value{

}
\description{
This function estimates parameters for one or more bias models
for a single sample over a set of single-isoform
genes. ~100 medium to highly expressed genes should be sufficient to
estimate the parameters robustly.
}

