% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bias.R
\name{fitModelOverGenes}
\alias{fitModelOverGenes}
\title{Fit bias model over single-isoform genes}
\usage{
fitModelOverGenes(genes, bamfile, fragtypes, genome, models, readlength,
  minsize, maxsize, zerotopos = 2, speedglm = TRUE)
}
\arguments{
\item{genes}{a GRangesList with the exons of different genes}

\item{bamfile}{a character string pointing to an indexed BAM file}

\item{fragtypes}{the output of \link{buildFragtypes}. must contain
the potential fragment types for the genes named in \code{genes}}

\item{genome}{a BSGenome object}

\item{models}{a list of character strings or formula describing the bias models, see vignette}

\item{readlength}{the read length}

\item{minsize}{the minimum fragment length to model}

\item{maxsize}{the maximum fragment length to model}

\item{zerotopos}{the ratio of zero counts (unobserved fragments) to positive counts
(observed fragments) that should be used in the Poisson GLM. The default value is 2.
The unobserved fragment types are then downsampled to match this ratio, and
up-weighted in the Poisson GLM.}

\item{speedglm}{logical, whether to use speedglm to estimate the coefficients.
Default is TRUE.}
}
\value{

}
\description{
This function estimates bias parameters for a single sample
over a set of single-isoform
genes. ~100 medium to highly expressed genes should be sufficient to
estimate the bias parameters robustly.
}

