% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getFragmentWidths}
\alias{getFragmentWidths}
\title{Get fragment widths}
\usage{
getFragmentWidths(bam.file, tx)
}
\arguments{
\item{bam.file}{a character string pointing to a BAM file}

\item{tx}{a GRanges object of the exons of a single isoform gene}
}
\value{
a numeric vector of estimated fragment widths
}
\description{
From a BAM file and a particular transcript (recommened
to be the single isoform of a gene), this function
returns estimates of the fragment widths, by mapping the
fragment alignments to the transcript coordinates.
}
\examples{

# these next lines just write out a BAM file from R
# typically you would already have a BAM file
library(alpineData)
library(GenomicAlignments)
library(rtracklayer)
gap <- ERR188088()
dir <- system.file(package="alpineData", "extdata")
bam.file <- c("ERR188088" = file.path(dir,"ERR188088.bam"))
export(gap, con=bam.file)

data(preprocessedData)

w <- getFragmentWidths(bam.file, ebt.fit[[2]])
quantile(w, c(.025, .975))

}

