% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getReadLength}
\alias{getReadLength}
\title{Get read length}
\usage{
getReadLength(bamfiles)
}
\arguments{
\item{bamfiles}{a character vector pointing to BAM files}
}
\value{
a numeric vector, one number per BAM file, the
length of the first read in the file
}
\description{
Gets the length of the first read in a BAM file
}

