% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{normalizeDESeq}
\alias{normalizeDESeq}
\title{DESeq median ratio normalization for matrix}
\usage{
normalizeDESeq(mat, cutoff)
}
\arguments{
\item{mat}{a matrix of numeric values}

\item{cutoff}{a numeric value to be used as the cutoff
for the row means of \code{mat}. Only rows with row mean
larger than \code{cutoff} are used for calculating
the size factors}
}
\value{
a matrix with the median ratio size factors
divided out
}
\description{
Simple implementation of DESeq median ratio normalization
}
\references{
Anders, S. and Huber, W.,
Differential expression analysis for sequence count data.
Genome Biology (2010) doi: 10.1186/gb-2010-11-10-r106
}

