% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFragLen}
\alias{plotFragLen}
\title{Plot fragment length distribution over samples}
\usage{
plotFragLen(fitpar, col, lty)
}
\arguments{
\item{fitpar}{a list of the output of \link{fitModelOverGenes} over samples}

\item{col}{a vector of colors}

\item{lty}{a vector of line types}
}
\description{
Plots the fragment length distribution.
}

