% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGC}
\alias{plotGC}
\title{Plot the fragment GC bias over samples}
\usage{
plotGC(fitpar, model, col, lty, ylim, knots = c(0.4, 0.5, 0.6), bk = c(0,
  1))
}
\arguments{
\item{fitpar}{a list of the output of \link{fitModelOverGenes} over samples}

\item{model}{the name of one of the models}

\item{col}{a vector of colors}

\item{lty}{a vector of line types}

\item{ylim}{the y limits for the plot}

\item{knots}{the knots for the spline}

\item{bk}{the boundary knots for the spline}
}
\description{
Plots smooth curves of the log fragment rate over fragment GC content.
}

