% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGRL}
\alias{plotGRL}
\title{Simple segments plot for GRangesList}
\usage{
plotGRL(grl, ...)
}
\arguments{
\item{grl}{GRangesList object}

\item{...}{passed to plot}
}
\description{
Simple segments plot for GRangesList
}

