% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOrder0}
\alias{plotOrder0}
\alias{plotOrder1}
\alias{plotOrder2}
\title{Plot parameters of the variable length Markov model (VLMM) for read starts}
\usage{
plotOrder0(order0, ...)

plotOrder1(order1, pos1)

plotOrder2(order2, pos2)
}
\arguments{
\item{order0}{the "order0" element of the list named "vlmm.fivep" or "vlmm.threep"
within the list that is the output of \link{fitBiasModels}}

\item{...}{parameters passed to \code{plot}}

\item{order1}{as for "order0" but "order1"}

\item{pos1}{the position of the 1st order VLMM to plot}

\item{order2}{as for "order0" but "order2"}

\item{pos2}{the position of the 2nd order VLMM to plot}
}
\description{
This function plots portions of the Cufflinks VLMM for read start bias.
The natural log of observed over expected is shown, such that 0
indicates no contribution of a position to the read start bias.
As the variable lenght Markov model has different dependencies for different
positions (see Roberts et al, 2011), it is difficult
to show all the 744 parameters simultaneously. Instead this function
offers to show the 0-order terms for all positions, or the 1st and 2nd
order terms for selected positions within the read start sequence.
For the 1- and 2-order terms, the log bias is shown for each nucleotide
(A,C,T,G) given the previous nucleotide (1-order) or di-nucleotide (2-order).
}
\section{Functions}{
\itemize{
\item \code{plotOrder1}: Plot first order parameters for a position

\item \code{plotOrder2}: Plot second order parameters for a position
}}
\examples{

# fitpar was fit using identical code
# as found in the vignette, except with
# 25 genes, and with fragment size in 80-350 bp
data(fitpar)
plotOrder0(fitpar[[1]][["vlmm.fivep"]][["order0"]])
plotOrder1(fitpar[[1]][["vlmm.fivep"]][["order1"]], pos1=5:19)
plotOrder2(fitpar[[1]][["vlmm.fivep"]][["order2"]], pos2=8:17)

}
\references{
Roberts et al, "Improving RNA-Seq expression estimates by correcting for fragment bias"
Genome Biology (2011) doi:101186/gb-2011-12-3-r22
}

