% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRelPos}
\alias{plotRelPos}
\title{Plot relative position bias over samples}
\usage{
plotRelPos(fitpar, model, col, lty, ylim)
}
\arguments{
\item{fitpar}{a list of the output of \link{fitBiasModels} over samples}

\item{model}{the name of one of the models}

\item{col}{a vector of colors}

\item{lty}{a vector of line types}

\item{ylim}{the y limits for the plot}
}
\value{
plot
}
\description{
Plots the smooth curves of log fragment rate over relative position.
}
\examples{

# fitpar was fit using identical code
# as found in the vignette, except with
# 25 genes, and with fragment size in 80-350 bp
data(preprocessedData)
perf <- rep(1:2, each=2)
plotRelPos(fitpar, "all", col=perf)

}

