% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{splitLongGenes}
\alias{splitLongGenes}
\title{Split very long genes}
\usage{
splitLongGenes(ebg, ebt, txdf, long = 1e+06)
}
\arguments{
\item{ebg}{an exons-by-genes GRangesList, created with \code{exonsBy}}

\item{ebt}{an exons-by-tx GRangesList, created with \code{exonsBy}}

\item{txdf}{a data.frame created by running \code{select} on a TxDb object.
Must have columns GENEID and TXNAME, where TXNAME corresponds to the
names of \code{ebt}. Note: this requires renaming \code{ebt}.}

\item{long}{a numeric value such that ranges longer than this are "long"}
}
\value{
a list of manipulated \code{ebg} and \code{txdf}
}
\description{
This function splits genes which have a very long range (e.g. 1 Mb),
and new "genes" are formed where each isoform is its own "gene",
with the suffix "_ls" and a number.
It makes sense to turn each isoform into its own gene only if this
function is followed by \link{mergeGenes}.
}

