% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extractAlpine}
\alias{extractAlpine}
\title{Extract results from estimateTheta run across genes}
\usage{
extractAlpine(res, model, nsamp, lib.sizes = 1e+06, divideOut = TRUE,
  transcripts = NULL)
}
\arguments{
\item{res}{a list where each element is the output of \link{estimateTheta}}

\item{model}{the name of a model, corresponds to names of \code{models}
used in \link{fitBiasModels}}

\item{nsamp}{the number of samples, corresponds to length of \code{bamfiles}
in \link{estimateTheta}}

\item{lib.sizes}{the vector of library sizes passed to \link{estimateTheta}.
not needed if \code{divideOut=FALSE}}

\item{divideOut}{logical, whether to divide out the initial estimate of
library size and to instead use the count of compatible fragments for
genes calculated by \link{estimateTheta}. Default is TRUE}

\item{transcripts}{an optional \code{GRangesList} of the exons for each
transcript. If this is provided, the output will be a
\code{SummarizedExperiment}. The transcripts do not need
to be provided in the correct order, \code{extractAlpine} will
find the correct transcript by the names in \code{res} and
put them in the correct order.}
}
\value{
a matrix of FPKM values across transcripts and samples
}
\description{
This function extracts estimates for a given model from a list
over many genes, returning a matrix with dimensions:
number of transcript x number of samples.
Here, the count of compatible fragments aligning to the
genes is used to estimate the FPKM, dividing out the previously
used estimate \code{lib.sizes}.
}

