% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bias.R
\name{fitBiasModels}
\alias{fitBiasModels}
\title{Fit bias models over single-isoform genes}
\usage{
fitBiasModels(genes, bam.file, fragtypes, genome, models, readlength, minsize,
  maxsize, speedglm = TRUE)
}
\arguments{
\item{genes}{a GRangesList with the exons of different
single-isoform genes}

\item{bam.file}{a character string pointing to an indexed BAM file}

\item{fragtypes}{the output of \link{buildFragtypes}. must contain
the potential fragment types for the genes named in \code{genes}}

\item{genome}{a BSgenome object}

\item{models}{a list of lists: the outer list describes multiple models
each element of the inner list has two elements: \code{formula} and \code{offset}.
\code{formula} should be a character strings of an R formula
describing the bias models, e.g. \code{"count ~ ns(gc) + gene"}.
\code{offset} should be a character vector
listing possible bias offsets to be used (\code{"fraglen"} or \code{"vlmm"}).
Either \code{offset} or \code{formula} can be \code{NULL} for a model.
See vignette for recommendations and details.}

\item{readlength}{the read length}

\item{minsize}{the minimum fragment length to model}

\item{maxsize}{the maximum fragment length to model}

\item{speedglm}{logical, whether to use speedglm to estimate the coefficients.
Default is TRUE.}
}
\value{
a list with elements: coefs, summary, models,
and optional offets: fraglen.density, vlmm.fivep,
and vlmm.threep. \strong{coefs} gives the estimated coefficients
for the different models that specified formula.
\strong{summary} gives the tables with coefficients, standard
errors and p-values, \strong{models} stores the incoming
\code{models} list, \strong{fraglen.density} is a
estimated density object for the fragment length distribution,
and \strong{vlmm.fivep} and \strong{vlmm.threep}
store the observed and expected tabulations for the different
orders of the VLMM for read start bias.
}
\description{
This function estimates parameters for one or more bias models
for a single sample over a set of single-isoform
genes. ~100 medium to highly expressed genes should be sufficient to
estimate the parameters robustly.
}
\examples{

# see vignette for a more realistic example

library(GenomicRanges)
library(BSgenome.Hsapiens.NCBI.GRCh38)
dir <- system.file("data",package="alpine")
load(file.path(dir,"ebtfit.rda"))
readlength <- 100
minsize <- 125 # see vignette how to choose
maxsize <- 175 # see vignette how to choose
# here a very small subset, should be ~100 genes
gene.names <- names(ebt.fit)[6:8] 
names(gene.names) <- gene.names
fragtypes <- lapply(gene.names, function(gene.name) {
                      buildFragtypes(ebt.fit[[gene.name]],
                                     Hsapiens, readlength,
                                     minsize, maxsize)
})
models <- list(
  "GC" = list(formula = "count ~ ns(gc,knots=gc.knots,
                Boundary.knots=gc.bk) + gene",
              offset=c("fraglen"))
)
bam.file <- "~/proj/alpine/alpine/inst/extdata/ERR188088_galignpairs.bam"
fitpar <- fitBiasModels(genes=ebt.fit[gene.names],
                        bam.file=bam.file,
                        fragtypes=fragtypes,
                        genome=Hsapiens,
                        models=models,
                        readlength=readlength,
                        minsize=minsize,
                        maxsize=maxsize)

}

