% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bias.R
\name{fitBiasModels}
\alias{fitBiasModels}
\title{Fit bias models over single-isoform genes}
\usage{
fitBiasModels(genes, bam.file, fragtypes, genome, models, readlength, minsize,
  maxsize, speedglm = TRUE)
}
\arguments{
\item{genes}{a GRangesList with the exons of different
single-isoform genes}

\item{bam.file}{a character string pointing to an indexed BAM file}

\item{fragtypes}{the output of \link{buildFragtypes}. must contain
the potential fragment types for the genes named in \code{genes}}

\item{genome}{a BSgenome object}

\item{models}{a list of lists: the outer list describes multiple models
each element of the inner list has two elements: \code{formula} and \code{offset}.
\code{formula} should be a character strings of an R formula
describing the bias models, e.g. \code{"count ~ ns(gc) + gene"}.
\code{offset} should be a character vector
listing possible bias offsets to be used (\code{"fraglen"} or \code{"vlmm"}).
Either \code{offset} or \code{formula} can be \code{NULL} for a model.
See vignette for recommendations and details.}

\item{readlength}{the read length}

\item{minsize}{the minimum fragment length to model}

\item{maxsize}{the maximum fragment length to model}

\item{speedglm}{logical, whether to use speedglm to estimate the coefficients.
Default is TRUE.}
}
\value{
TODO a list containing bias parameters
}
\description{
This function estimates parameters for one or more bias models
for a single sample over a set of single-isoform
genes. ~100 medium to highly expressed genes should be sufficient to
estimate the parameters robustly.
}

