% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getFragmentWidths}
\alias{getFragmentWidths}
\title{Get fragment widths}
\usage{
getFragmentWidths(bamfile, tx)
}
\arguments{
\item{bamfile}{a character string pointing to a BAM file}

\item{tx}{a GRanges object of the exons of a single isoform gene}
}
\value{
a numeric vector of estimated fragment widths
}
\description{
From a BAM file and a particular transcript (recommened
to be the single isoform of a gene), this function
returns estimates of the fragment widths, by mapping the
fragment alignments to the transcript coordinates.
}

