% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getFragmentWidths}
\alias{getFragmentWidths}
\title{Get fragment widths}
\usage{
getFragmentWidths(bam.file, tx)
}
\arguments{
\item{bam.file}{a character string pointing to a BAM file}

\item{tx}{a GRanges object of the exons of a single isoform gene}
}
\value{
a numeric vector of estimated fragment widths
}
\description{
From a BAM file and a particular transcript (recommened
to be the single isoform of a gene), this function
returns estimates of the fragment widths, by mapping the
fragment alignments to the transcript coordinates.
}
\examples{

data(ebtfit)
bam.file <- "~/proj/alpine/alpine/inst/extdata/ERR188088_galignpairs.bam"
w <- getFragmentWidths(bam.file, ebt.fit[[2]])
quantile(w, c(.025, .975))

}

