% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mergeGenes}
\alias{mergeGenes}
\title{Merge overlapping "genes" into gene clusters}
\usage{
mergeGenes(ebg, txdf, ignore.strand = TRUE)
}
\arguments{
\item{ebg}{an exons-by-genes GRangesList, created with \code{exonsBy}}

\item{txdf}{a data.frame created by running \code{select} on a TxDb object.
Must have a column GENEID.}

\item{ignore.strand}{Default is TRUE.}
}
\value{
a manipulated \code{txdf}.
}
\description{
This function looks for overlapping exons in \code{ebg}.
The overlapping "genes" are used to form a graph.
Any connected components in the graph (sets of "genes"
which can be reached from each other through overlap relations)
are connected into a new gene cluster, which is given the
suffix "_mrg" and using one of the original gene names.
}

