% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotFragLen}
\alias{plotFragLen}
\title{Plot fragment length distribution over samples}
\usage{
plotFragLen(fitpar, col, lty)
}
\arguments{
\item{fitpar}{a list of the output of \link{fitBiasModels} over samples}

\item{col}{a vector of colors}

\item{lty}{a vector of line types}
}
\description{
Plots the fragment length distribution.
}
\examples{

# fitpar was fit using identical code
# as found in the vignette, except with
# 25 genes, and with fragment size in 80-350 bp
data(fitpar)
perf <- rep(1:2, each=2)
plotFragLen(fitpar, col=perf)

}

