% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGC}
\alias{plotGC}
\title{Plot the fragment GC bias over samples}
\usage{
plotGC(fitpar, model, col, lty, ylim, knots = c(0.4, 0.5, 0.6), bk = c(0,
  1), gc.range = NULL, return.type = 0)
}
\arguments{
\item{fitpar}{a list of the output of \link{fitBiasModels} over samples}

\item{model}{the name of one of the models}

\item{col}{a vector of colors}

\item{lty}{a vector of line types}

\item{ylim}{the y limits for the plot}

\item{knots}{the knots for the spline}

\item{bk}{the boundary knots for the spline}

\item{gc.range}{a numeric of length two,
the range of the fragment GC content. By default,
[.2,.8] for plotting and [0,1] for returning a matrix}

\item{return.type}{a numeric, either
0: make a plot,
1: skip the plot and return a matrix of log fragment rate,
2: skip the plot and return a matrix of probabilities}
}
\value{
Either plot, or if \code{return.type} is 1 or 2, a matrix
}
\description{
Plots smooth curves of the log fragment rate over fragment GC content.
}
\examples{

# fitpar was fit using identical code
# as found in the vignette, except with
# 25 genes, and with fragment size in 80-350 bp
data(preprocessedData)
perf <- rep(1:2, each=2)
plotGC(fitpar, "all", col=perf)

}

