% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGRL}
\alias{plotGRL}
\title{Simple segments plot for GRangesList}
\usage{
plotGRL(grl, ...)
}
\arguments{
\item{grl}{GRangesList object}

\item{...}{passed to plot}
}
\value{
plot
}
\description{
Simple segments plot for GRangesList
}
\examples{

library(GenomicRanges)
grl <- GRangesList(GRanges("1",IRanges(c(100,200,300),width=50)),
                   GRanges("1",IRanges(c(100,300),width=c(75,50))),
                   GRanges("1",IRanges(c(100,200,400),width=c(75,50,50))),
                   GRanges("1",IRanges(c(200,300,400),width=50)))
plotGRL(grl)

}

