% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRelPos}
\alias{plotRelPos}
\title{Plot relative position bias over samples}
\usage{
plotRelPos(fitpar, model, col, lty, ylim, knots = c(0.25, 0.5, 0.75),
  bk = c(0, 1))
}
\arguments{
\item{fitpar}{a list of the output of \link{fitBiasModels} over samples}

\item{model}{the name of one of the models}

\item{col}{a vector of colors}

\item{lty}{a vector of line types}

\item{ylim}{the y limits for the plot}

\item{knots}{the knots for the spline}

\item{bk}{the boundary knots for the spline}
}
\value{
plot
}
\description{
Plots the smooth curves of log fragment rate over relative position.
}
\examples{

# fitpar was fit using identical code
# as found in the vignette, except with
# 25 genes, and with fragment size in 80-350 bp
data(preprocessedData)
perf <- rep(1:2, each=2)
plotRelPos(fitpar, "all", col=perf)

}

