% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{splitGenesAcrossChroms}
\alias{splitGenesAcrossChroms}
\title{Split genes that have isoforms across chromosomes}
\usage{
splitGenesAcrossChroms(ebg, txdf)
}
\arguments{
\item{ebg}{an exons-by-genes GRangesList, created with \code{exonsBy}}

\item{txdf}{a data.frame created by running \code{select} on a TxDb object.
Must have columns TXCHROM and GENEID}
}
\value{
a list of manipulated \code{ebg} and \code{txdf}
}
\description{
This function simply splits apart genes which have isoforms across multiple
chromosomes. New "genes" are created with the suffix "_cs" and a number.
}

